core/requirements.txt
InquirerPy~=0.3.4
typer~=0.17.4
dotenv~=0.9.9
pydantic~=2.11.7
fastapi~=0.116.1
uvicorn~=0.35.0
pydbml~=1.2.0
inflect~=7.5.0
jinja2~=3.1.6
strands-agents
pyparsing~=3.2.3
redis~=5.0.1
